notification-service/requirements.txt
apifairy==1.3.0
apispec==6.3.0
Flask==3.0.0
flask-marshmallow==0.15.0
Flask-SocketIO==5.3.6
marshmallow==3.20.1
paho-mqtt==1.5.1
marshmallow-enum==1.5.1
pymongo==4.6.1
Flask-Mail==0.9.1